Rounded out the SavedSearch model. The save() method now calls _make_qs to generate each search substring, then after each of those _make_qs calls complete, it calls _full_qs to generate the self.querystring attribute, which is what will ultimately be passed to Solr.
I made an implementation decision to bypass the Haystack API and make each SavedSearch model have that querystring attribute, so that it can be passed directly to the address specified in settings.HAYSTACK_SOLR_URL. The other choice was to have SavedSearch instances have a SearchQuerySet object exposed as an attribute, which may turn out to be necessary soon. I'll jump off that bridge when I come to it, though.