Improve handling of multi-term attributes, such as multiple titles, keywords and cities in a single SavedSearch querystring.
This was handled by converting the values of the attribute dict returned by _attr_dict method from strings to lists. The elements of those lists are reduced to a single SQ. Each of those SQs (for a given SavedSearch instnace) are then reduced to a single SQ object and passed to the search engine to run a single query of any given complexity.